Adding discovery loop to find peers
Added a discover method (infinite loop on separate thread) that continuously sends small messages (heartbeats) to the other users on the same port. This will allow for the different users to recognize each other and whatnot, honestly probably a wasted implementation at this point but we ball regardless 😎. Next, I plan on restructuring the repo to be based off of UDP multicast instead of UDP broadcast, as multicast works better on windows and has overall more flexibility. It is also more widely used in industry.