Converting to UDP Multicast
Restructuring to use UDP multicast instead of UDP broadcast, which allows for more expandability (multiple port usage) and is more industry-like. It also makes it work so that's pretty nice. Going to fix up some of the print statements to make it a little cleaner (not too important, as GUI will be implemented later), then start working on the wrapper functions for signatures and encryption/decryption.