Combining redundant functions
Fixing up the functionality a little bit by cleaning up redundant functions in crypto_utils. I was going to retrieve the cyphertext and signature by hand, however this will likely be too complex for now so I will use pickle instead. I also hope to add keys to the beginning of each message so the listener knows what to do with it. This could also (potentially) make the GUI implementation a little easier as well as making the code more interpretable.